+++ src/context/testCaseActionTypes.js
@@ -1,0 +1,53 @@
+export const actionTypes = {
+  UPDATE_STATEMENTS_ORDER: 'UPDATE_STATEMENTS_ORDER',
+  UPDATE_TEST_STATEMENT: 'UPDATE_TEST_STATEMENT',
+
+  ADD_ACTION: 'ADD_ACTION',
+  DELETE_ACTION: 'DELETE_ACTION',
+  UPDATE_ACTION: 'UPDATE_ACTION',
+
+  ADD_MIDDLEWARE: 'ADD_MIDDLEWARE',
+  DELETE_MIDDLEWARE: 'DELETE_MIDDLEWARE',
+  UPDATE_MIDDLEWARE: 'UPDATE_MIDDLEWARE',
+
+  ADD_ASSERTION: 'ADD_ASSERTION',
+  DELETE_ASSERTION: 'DELETE_ASSERTION',
+  UPDATE_ASSERTION: 'UPDATE_ASSERTION',
+
+  ADD_RENDER: 'ADD_RENDER',
+  DELETE_RENDER: 'DELETE_RENDER',
+  UPDATE_RENDER_COMPONENT: 'UPDATE_RENDER_COMPONENT',
+
+  ADD_RENDER_PROP: 'ADD_RENDER_PROP',
+  DELETE_RENDER_PROP: 'DELETE_RENDER_PROP',
+  UPDATE_RENDER_PROP: 'UPDATE_RENDER_PROPS',
+
+  ADD_ASYNC: 'ADD_ASYNC',
+  DELETE_ASYNC: 'DELETE_ASYNC',
+  UPDATE_ASYNC: 'UPDATE_ASYNC',
+
+  ADD_ACTIONCREATOR: 'ADD_ACTIONCREATOR',
+  DELETE_ACTIONCREATOR: 'DELETE_ACTIONCREATOR',
+  UPDATE_ACTIONCREATOR: 'UPDATE_ACTIONCREATORS',
+
+  ADD_REDUCER: 'ADD_REDUCER',
+  DELETE_REDUCER: 'DELETE_REDUCER',
+  UPDATE_REDUCER: 'UPDATE_REDUCER',
+
+  // Action type for hookRender
+  ADD_HOOKRENDER: 'ADD_HOOKRENDER',
+  DELETE_HOOKRENDER: 'DELETE_HOOKRENDER',
+  UPDATE_HOOKRENDER: 'UPDATE_HOOKRENDER',
+
+  ADD_HOOK_UPDATES: 'ADD_HOOK_UPDATES',
+  DELETE_HOOK_UPDATES: 'DELETE_HOOK_UPDATE',
+  UPDATE_HOOK_UPDATES: 'UPDATE_HOOK_UPDATES',
+
+  UPDATE_ACTIONS_FILEPATH: 'UPDATE_ACTIONS_FILEPATH',
+  UPDATE_TYPES_FILEPATH: 'UPDATE_TYPES_FILEPATH',
+  UPDATE_REDUCERS_FILEPATH: 'UPDATE_REDUCERS_FILEPATH',
+  UPDATE_MIDDLEWARES_FILEPATH: 'UPDATE_MIDDLEWARES_FILEPATH',
+  UPDATE_HOOKS_FILEPATH: 'UPDATE_HOOKS_FILEPATH',
+
+  CREATE_NEW_TEST: 'CREATE_NEW_TEST',
+};
